pyproject.toml
[project]
name = "sanskrit-vi-translation"
version = "0.1.0"
description = "Sanskrit-Vietnamese Translation Benchmark"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bert-score>=0.3.13",
    "litellm>=1.81.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sacrebleu>=2.6.0",
    "torch>=2.9.1",
    "requests",
    "beautifulsoup4",
    "cloudscraper",
    "sacremoses",
    "tabulate",
    # Langfuse Observability
    "langfuse>=3.0.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-exporter-otlp>=1.39.1",
    "hydra-core>=1.3.2",
    "diskcache>=5.6.3",
    "pydantic>=2.12.5",
    "tenacity>=9.1.2",
    "jinja2>=3.1.6",
]

[project.optional-dependencies]
vllm = ["vllm>=0.6.0"]
